requirements.txt
numpy
opencv-python
mediapipe
pillow
